Add .NET Core 3.0 support to projects. Also applied axe. Doesn't work
