+++ Generator/ConfigurationKeys.template.cs
