--- NFluidsynth.MidiAccess.Shared/NFluidsynth.MidiAccess.Shared.projitems
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>{72F17E3C-B5C0-4E75-BBAD-B51FF560CCFD}</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>NFluidsynth.MidiAccess.Shared</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)FluidsynthMidiAccess.cs" />
-  </ItemGroup>
-</Project>
