+++ NFluidsynth.MidiAccess/NFluidsynth.MidiAccess.csproj
@@ -1,54 +1,16 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{1C0A7DD2-6F9A-457A-A70C-7BF23E5E6D2C}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <RootNamespace>NFluidsynth.MidiAccess</RootNamespace>
-    <AssemblyName>NFluidsynth.MidiAccess</AssemblyName>
-    <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug</OutputPath>
-    <DefineConstants>DEBUG;</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <ConsolePause>false</ConsolePause>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release</OutputPath>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <ConsolePause>false</ConsolePause>
+﻿<Project Sdk="Microsoft.NET.Sdk">
+<PropertyGroup>
+    <TargetFrameworks>net461;netcoreapp3.0</TargetFrameworks>
+    <Copyright>atsushi</Copyright>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="alsa-sharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null">
-      <HintPath>..\packages\managed-midi.1.9.0\lib\net45\alsa-sharp.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
-    <Reference Include="Commons.Music.Midi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null">
-      <HintPath>..\packages\managed-midi.1.9.0\lib\net45\Commons.Music.Midi.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
-    <Reference Include="System, Version=1.0.5000.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
+    <PackageReference Include="managed-midi" Version="1.9.0" />
+    <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.0-preview.2">
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+      <PrivateAssets>all</PrivateAssets>
+    </PackageReference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Properties\AssemblyInfo.cs" />
+    <ProjectReference Include="..\NFluidsynth\NFluidsynth.csproj" />
   </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\NFluidsynth\NFluidsynth.Desktop.csproj">
-      <Project>{79326AFF-EA30-479B-A48E-77CF5AB99EF4}</Project>
-      <Name>NFluidsynth.Desktop</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="packages.config" />
-  </ItemGroup>
-  <Import Project="..\NFluidsynth.MidiAccess.Shared\NFluidsynth.MidiAccess.Shared.projitems" Label="Shared" Condition="Exists('..\NFluidsynth.MidiAccess.Shared\NFluidsynth.MidiAccess.Shared.projitems')" />
-  <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
 </Project>
