--- NFluidsynth.MidiAccess/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="managed-midi" version="1.9.0" targetFramework="net45" />
-</packages>
