--- NFluidsynth/generator/.gitignore
@@ -1,2 +1,0 @@
-proplistgen.exe
-proplistgen.pdb
